+++ Assets/_Project/Scripts/DeathOfShip.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 62709447223e0664cac599d310eafa63
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
