+++ Assets/_Project/Scripts/ShipMovement.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2f40dce4b87a4b84e9901a9cd072c1fb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
